添加canvas支持动画的功能

+++ GICXMLLayout/Classes/Core/core/GICLayoutElementProtocol.h
@@ -44,7 +44,7 @@
  */
 -(id)gic_addSubElement:(id)subElement;
 
--(id)gic_insertSubElement:(id)subElement elementOrder:(NSInteger)order;
+-(id)gic_insertSubElement:(id)subElement elementOrder:(CGFloat)order;
 
 
 /**
